+++ packages/project-config/src/index.ts
@@ -30,7 +30,7 @@ const logger = logging.getLogger('polymer-project-config');
  */
 export const defaultSourceGlobs = ['src/**/*'];
 
-export type ModuleResolutionStrategy = 'none' | 'node';
+export type ModuleResolutionStrategy = 'none'|'node';
 const moduleResolutionStrategies =
     new Set<ModuleResolutionStrategy>(['none', 'node']);
 
@@ -325,7 +325,7 @@ export class ProjectConfig {
     /**
      * moduleResolution
      */
-    this.moduleResolution = options.moduleResolution || 'none';
+    this.moduleResolution = options.moduleResolution || 'node';
 
     /**
      * root