+++ packages/project-config/CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-<!-- ## Unreleased -->
+## Unreleased
+* Change the default value of --module-resolution to "node".
 <!-- Add new, unreleased changes here. -->
 
 ## [3.14.0] - 2018-05-03